[ACS-7965] Remove duplications
